+++ lib/connect.h
@@ -111,7 +111,7 @@ void Curl_conncontrol(struct connectdata *conn,
 
 CURLcode Curl_cf_setup_insert_after(struct Curl_cfilter *cf_at,
                                     struct Curl_easy *data,
-                                    int transport,
+                                    uint8_t transport,
                                     int ssl_mode);
 
 /**
